requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
Django==4.2.3
django-bootstrap-v5==1.0.11
django-cors-headers==4.2.0
django-extensions==3.2.3
djangorestframework==3.14.0
pytz==2023.3
soupsieve==2.4.1
sqlparse==0.4.4
typing_extensions==4.7.1
